+++ Enums/BotType.cs
@@ -1,0 +1,8 @@
+namespace WeatherMonitoring.WeatherBots;
+
+public enum BotType
+{
+    RainBot = 0,
+    SunBot = 1,
+    SnowBot = 2
+}
