--- IWeatherBots/BotType.cs
@@ -1,8 +1,0 @@
-namespace WeatherMonitoring.IWeatherBots;
-
-public enum BotType
-{
-    RainBot,
-    SunBot,
-    SnowBot
-}
