+++ Program.cs
@@ -1,9 +1,9 @@
-﻿namespace WeatherMonitoring.IWeatherBots;
+﻿namespace WeatherMonitoring.WeatherBots;
 
 public class Program
 {
     private const string ConfigFileName = "Configuration.json";
-    
+
     public static void Main()
     {
         var configs = ConfigLoader.Load(ConfigFileName);
