+++ WeatherBots/BotConfigurations/BotConfiguration.cs
@@ -1,3 +1,3 @@
-namespace WeatherMonitoring.IWeatherBots.BotConfigurations;
+namespace WeatherMonitoring.WeatherBots.BotConfigurations;
 
 public record BotConfiguration(string Message, bool Enabled);
