+++ WeatherBots/BotConfigurations/TemperatureConfiguration.cs
@@ -1,3 +1,3 @@
-namespace WeatherMonitoring.IWeatherBots.BotConfigurations;
+namespace WeatherMonitoring.WeatherBots.BotConfigurations;
 
 public record TemperatureConfiguration(double Treshold, string Message, bool Enabled) : BotConfiguration(Message, Enabled);
