+++ WeatherBots/ConfigLoader.cs
@@ -1,7 +1,7 @@
 using System.Text.Json;
-using  WeatherMonitoring.IWeatherBots.BotConfigurations;
+using WeatherMonitoring.WeatherBots.BotConfigurations;
 
-namespace WeatherMonitoring.IWeatherBots;
+namespace WeatherMonitoring.WeatherBots;
 
 public static class ConfigLoader
 {
